Rename "oneshot" to be more descriptive: "stateless"
