+++ .gitignore
@@ -8,8 +8,8 @@ checksums.txt
 crash-*
 index.cjs
 index.d.cts
-oneshot.cjs
-oneshot.d.cts
+stateless.cjs
+stateless.d.cts
 testing.cjs
 testing.d.cts
