+++ config/eslint.yml
@@ -755,7 +755,7 @@ ignorePatterns:
   - script/maybe-run.js
   - index.cjs
   - index.d.cts
-  - oneshot.cjs
-  - oneshot.d.cts
+  - stateless.cjs
+  - stateless.d.cts
   - testing.cjs
   - testing.d.cts
