+++ script/create-d-cts.js
@@ -8,7 +8,7 @@ import path from "node:path";
 
 import { common } from "./_.js";
 
-const files = ["index.d.ts", "oneshot.d.ts", "testing.d.ts"];
+const files = ["index.d.ts", "stateless.d.ts", "testing.d.ts"];
 
 for (const file of files) {
   const copy = file.replace(".d.ts", ".d.cts");
